added DynamicClassCompiler
SpoonClassLoader uses a JDTByteCompiler, used by Eclipse,
but some classes can not be compiled with these compiler
even when they are compiled successfully by JavaCompiler.
in this commit a DynamicClassCompiler is introduced to
compile classes in runtime from a String object.
the idea is to integrate this class in SpoonClassLoader.