fix bug in RuntimeValuesProcessor
- when a field of an anonymous class is reached from a statement
and the RuntimeValuesProcessor wants to add a statement to
collect the value of that field during runtime, the field name
cannot be "ClassName.this.nameOfField" because the name of the
class is empty. this was leading to compilation errors because
the name of the field being collected was ".this.nameOfField".
it is now corrected to "this.nameOfField"